requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-classic
google-generativeai

faiss-cpu
rank-bm25

PyPDF2
python-dotenv
pandas
python-docx

pinecone
langchain-pinecone